// src/types/postman-2.1.types.ts
export type NameOfTheCollection = string;
/**
 * A Description can be a raw text, or be an object, which holds the description along with its format.
 */
export type DefinitionsDescription = Description | string | null;
/**
 * Postman allows you to version your collections as they grow, and this field holds the version number. While optional, it is recommended that you use this field to its fullest extent!
 */
export type CollectionVersion =
  | {
      /**
       * Increment this number if you make changes to the collection that changes its behaviour. E.g: Removing or adding new test scripts. (partly or completely).
       */
      major: number;
      /**
       * You should increment this number if you make changes that will not break anything that uses the collection. E.g: removing a folder.
       */
      minor: number;
      /**
       * Ideally, minor changes to a collection should result in the increment of this number.
       */
      patch: number;
      /**
       * A human friendly identifier to make sense of the version numbers. E.g: 'beta-3'
       */
      identifier?: string;
      meta?: unknown;
      [k: string]: unknown;
    }
  | string;
export type Items = Item | Folder;
/**
 * Using variables in your Postman requests eliminates the need to duplicate requests, which can save a lot of time. Variables can be defined, and referenced to from any part of a request.
 */
export type Variable = Variable1 & Variable2;
export type Variable1 =
  | {
      [k: string]: unknown;
    }
  | {
      [k: string]: unknown;
    }
  | {
      [k: string]: unknown;
    };
/**
 * Collection variables allow you to define a set of variables, that are a *part of the collection*, as opposed to environments, which are separate entities.
 * *Note: Collection variables must not contain any sensitive information.*
 */
export type VariableList = Variable[];
// ...
export type EventList = Event[];
/**
 * A request represents an HTTP request. If a string, the string is assumed to be the request URL and the method is assumed to be 'GET'.
 */
export type Request = Request1 | string;
/**
 * The attributes for API Key Authentication.
 */
export type APIKeyAuthentication = Auth1[];
/**
 * The attributes for [AWS Auth](http://docs.aws.amazon.com/AmazonS3/latest/dev/RESTAuthentication.html).
 */
export type AWSSignatureV4 = Auth1[];
/**
 * The attributes for [Basic Authentication](https://en.wikipedia.org/wiki/Basic_access_authentication).
 */
export type BasicAuthentication = Auth1[];
/**
 * The helper attributes for [Bearer Token Authentication](https://tools.ietf.org/html/rfc6750)
 */
export type BearerTokenAuthentication = Auth1[];
/**
 * The attributes for [Digest Authentication](https://en.wikipedia.org/wiki/Digest_access_authentication).
 */
export type DigestAuthentication = Auth1[];
/**
 * The attributes for [Akamai EdgeGrid Authentication](https://developer.akamai.com/legacy/introduction/Client_Auth.html).
 */
export type EdgeGridAuthentication = Auth1[];
/**
 * The attributes for [Hawk Authentication](https://github.com/hueniverse/hawk)
 */
export type HawkAuthentication = Auth1[];
/**
 * The attributes for [NTLM Authentication](https://msdn.microsoft.com/en-us/library/cc237488.aspx)
 */
export type NTLMAuthentication = Auth1[];
/**
 * The attributes for [OAuth2](https://oauth.net/1/)
 */
export type OAuth1 = Auth1[];
/**
 * Helper attributes for [OAuth2](https://oauth.net/2/)
 */
export type OAuth2 = Auth1[];
// ...
export interface HttpsSchemaGetpostmanComJsonCollectionV210 {
  info: Information;
  /**
   * Items are the basic unit for a Postman collection. You can think of them as corresponding to a single API endpoint. Each Item has one request and may have multiple API responses associated with it.
   */
  item: Items[];
  event?: EventList;
  variable?: VariableList;
  auth?: null | Auth;
  protocolProfileBehavior?: ProtocolProfileBehavior;
  [k: string]: unknown;
}
/**
 * Detailed description of the info block
 */
export interface Information {
  name: NameOfTheCollection;
  /**
   * Every collection is identified by the unique value of this field. The value of this field is usually easiest to generate using a UID generator function. If you already have a collection, it is recommended that you maintain the same id since changing the id usually implies that is a different collection than it was originally.
   *  *Note: This field exists for compatibility reasons with Collection Format V1.*
   */
  _postman_id?: string;
  description?: DefinitionsDescription;
  version?: CollectionVersion;
  /**
   * This should ideally hold a link to the Postman schema that is used to validate this collection. E.g: https://schema.getpostman.com/collection/v1
   */
  schema: string;
  [k: string]: unknown;
}
export interface Description {
  /**
   * The content of the description goes here, as a raw string.
   */
  content?: string;
  /**
   * Holds the mime type of the raw description content. E.g: 'text/markdown' or 'text/html'.
   * The type is used to correctly render the description when generating documentation, or in the Postman app.
   */
  type?: string;
  /**
   * Description can have versions associated with it, which should be put in this property.
   */
  version?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
/**
 * Items are entities which contain an actual HTTP request, and sample responses attached to it.
 */
export interface Item {
  /**
   * A unique ID that is used to identify collections internally
   */
  id?: string;
  /**
   * A human readable identifier for the current item.
   */
  name?: string;
  description?: DefinitionsDescription;
  variable?: VariableList;
  event?: EventList;
  request: Request;
  response?: Responses;
  protocolProfileBehavior?: ProtocolProfileBehavior;
  [k: string]: unknown;
}
export interface Variable2 {
  /**
   * A variable ID is a unique user-defined value that identifies the variable within a collection. In traditional terms, this would be a variable name.
   */
  id?: string;
  /**
   * A variable key is a human friendly value that identifies the variable within a collection. In traditional terms, this would be a variable name.
   */
  key?: string;
  /**
   * The value that a variable holds in this collection. Ultimately, the variables will be replaced by this value, when say running a set of requests from a collection
   */
  value?: {
    [k: string]: unknown;
  };
  /**
   * A variable may have multiple types. This field specifies the type of the variable.
   */
  type?: 'string' | 'boolean' | 'any' | 'number';
  /**
   * Variable name
   */
  name?: string;
  description?: DefinitionsDescription;
  /**
   * When set to true, indicates that this variable has been set by Postman
   */
  system?: boolean;
  disabled?: boolean;
  [k: string]: unknown;
}
/**
 * Defines a script associated with an associated event name
 */
export interface Event {
  /**
   * A unique identifier for the enclosing event.
   */
  id?: string;
  /**
   * Can be set to `test` or `prerequest` for test scripts or pre-request scripts respectively.
   */
  listen: string;
  script?: Script;
  /**
   * Indicates whether the event is disabled. If absent, the event is assumed to be enabled.
   */
  disabled?: boolean;
  [k: string]: unknown;
}
// ...
export interface Auth {
  type:
    | 'apikey'
    | 'awsv4'
    | 'basic'
    | 'bearer'
    | 'digest'
    | 'edgegrid'
    | 'hawk'
    | 'noauth'
    | 'oauth1'
    | 'oauth2'
    | 'ntlm';
  noauth?: unknown;
  apikey?: APIKeyAuthentication;
  awsv4?: AWSSignatureV4;
  basic?: BasicAuthentication;
  bearer?: BearerTokenAuthentication;
  digest?: DigestAuthentication;
  edgegrid?: EdgeGridAuthentication;
  hawk?: HawkAuthentication;
  ntlm?: NTLMAuthentication;
  oauth1?: OAuth1;
  oauth2?: OAuth2;
  [k: string]: unknown;
}
// ...
export interface ProtocolProfileBehavior {
  [k: string]: unknown;
}
/**
 * One of the primary goals of Postman is to organize the development of APIs. To this end, it is necessary to be able to group requests together. This can be achieved using 'Folders'. A folder just is an ordered set of requests.
 */
export interface Folder {
  /**
   * A folder's friendly name is defined by this field. You would want to set this field to a value that would allow you to easily identify this folder.
   */
  name?: string;
  description?: DefinitionsDescription;
  variable?: VariableList;
  /**
   * Items are entities which contain an actual HTTP request, and sample responses attached to it. Folders may contain many items.
   */
  item: Items1[];
  event?: EventList;
  auth?: null | Auth;
  protocolProfileBehavior?: ProtocolProfileBehavior;
  [k: string]: unknown;
}